Update integrate-github-and-discord-with-webhooks.md
Remember how I was telling you your description shows up at the top of the post? Well it isn't happening for this one, which is weird. My hypothesis is that the quotes around your description might be the issue, so I've removed those (I wasn't able to reproduce the problem, but in my blog I found that the text shows up without the quotes whether it has them in the markdown or not). The other possible issue is the line break in the description (which I've also removed). When I tested that with my blog and ran it locally, I found that the page didn't show up at all, which is a completely different problem that doesn't seem to be happening with yours. Still, seems like something worth fixing, and it might solve the problem you've got.

You also had some weird captions on a couple images (in the form of text below them, not built into the image element), so I removed those.