Update excel-cumulative-flow-diagram.md
I changed your title image because you had two images at the top of the post and it looked wonky. The second one was cleaner and the first one (original "featured image" is shown later in the post, so I just moved that second one to be the "featured image." Other changes typo fixes.